requirements.txt
streamlit
streamlit-mic-recorder
openai-whisper
psycopg2-binary
sqlalchemy